Prevent Branch#base_ref being clobbered
Only the first push to a branch (i.e. the push that creates the branch)
has `base_ref` set, subsequent pushes have it as nil. This meant that
pushes after the first one were nil-ing out the value from the first
push.

Now we only set Branch#base_ref if we create a new branch - it may still
be nil (for branches that were initially pushed before we received or
processesed the relevant webhooks for the repo), but we avoid
overriding actual values with nil.