Scope next actions on boards#show to current user
We now work out all possible actions on a pull request, and tag them
with the users that they concern. For the PR dashboard, we take the top
priority item from this collection, and on boards#show we filter before-
hand to just the actions for that user, then show the top one.

Calculating all possible actions for a PR means we could cache this info
across the entire team, then do a quick filter per-user.