Ensure that blocking ticket actions are always flagged for users
Previously, if a blocking action didn't concern the user in question, it
was ignored and thus gave the false impression (subject to the other
action classes) that the PR could be merged.

We now ensure that, for every blocking class, we return an action for
all users. Most of the time, it's just "wait for the PR owner to fix
the blocker", but it means that problems are highlighted.