Add `permissive_enum`, which does not raise on invalid values
Instead, outputs a warning to the log and Bugsnag.

Covers the cases where GitHub adds a new state etc. value that appears
in webhooks - we want to be able to store these values without the
server erroring, whilst at the same time being notified so that we can
amend the app code accordingly.