Show time since last transition on card
Rather that *total* time in swimlane.

More efficient, since we can easily eager-load the corresponding model,
and potentially closer to what the user expects the number to represent.
We also given the number a much clearer label to clarify its meaning.

The total duration in the swimlane is still available in the ticket
modal.