Attempt fix for missing icons in production
Icon fonts can't found because they're not exported during
`assets:precompile`, so we copy them across after the precompile into
the location that the semantic-ui CSS expects them to be.

A kludgy fix, but a proper fix would mean moving to a Less-based setup,
and that's potentially a lot of work.