Fix NoMethodError when rebalancing BoardTickets
Caused by code in the ranked_model gem that uses `Array#index` to find
the current position of the ticket in the swimlane. Since the ticket
*isn't yet* in the swimlane, this returns `nil`, which triggers the
NoMethodError when the gem tries to call `#<` on said nil.

We work around this by putting the ticket into the desired swimlane in
the database, so that ranked_model picks it up correctly, but bypass the
model entirely so that it is still treated as a single change.