Upgrade Traefik to v2
Allows us to do SSL termination, and host+path routing, so that we can
fix HMR for webpacker (it's possible we could do this with v1, but the
docs are thin on the matter and we might as well be running the latest
version. YAML is also easier to read IMHO than TOML).

Some things to note:

- Traefik can infer the port to forward to from the exposed port on the
  container, so we don't need to explicitly set it.
- Priority is now calculated from the length of the rule, so again we
  don't need to set it explicitly.
- `traefik.http.routers.<container name>.tls = true` enables SSL
  termination.