Add script for running TailwindCSS
Currently, the version of PostCSS and NodeJS that we run are too old for
Tailwind v2, so we instead just download a newer version of NodeJS and
use a separate `package.json` to compile TailwindCSS, which can then be
pulled into webpack/Sprockets.

TODO: remove this once we get to webpacker v6 (or an earlier version
that supports PostCSS v7, as we can then use the `compat` build of
TailwindCSS).