Set up webpacker v6 for V2 UI
Only using the NPM part, as the gem requires Rails > 5.2, but that's the
more useful part (configuring webpack _right_ is hard). V6 is still in
beta, but a) allows us to have a separate `webpacker.yml` file, b)
supports PostCSS v8 (needed for TailwindCSS, although it 'supports' it
by not having a dependency on PostCSS at all) and c) doesn't require
copying in config files (Babel, PostCSS) like v5 does.

We crib the necessary packages, templates etc from the gem source:
https://github.com/rails/webpacker/blob/v6.0.0.beta.2/lib/install/template.rb
for the NPM packages & versions, and
https://github.com/rails/webpacker/tree/v6.0.0.beta.2/lib/install/config
for the config files. We don't have separate Webpack config files per
environment as we're only going be using it in dev and prod, and so can
set the necessary env vars in the wrapper scripts. We also need to
monkeypatch the webpacker package to stop it looking at the parent
node_modules folder, which it does to determine which loaders to include
(this fails, since the dependencies of those loaders have changed since
webpack v3).

Will write some small scripts to do the same job as the missing gem -
basically we need a way to read the `manifest.json` and inject the
relevant JS & CSS URLs into script/link tags, and a way to run webpack
during asset precompilation.