Refactor `mapStateToProps` to avoid prop-drilling
Instead, we pass around ids as props, and `connect()` components at a
lower level so that they can just get the info of the entity (and maybe
a few related entities as well). In most cases we know that the `id`
prop will never change for a component, so for these cases we use the
optimised version of `mapStateToProps` where we return a function that
closes over the `id` prop.

As a result, we replace the list components with components for the
individual items, since the wrappers weren't doing much (certainly their
interface was a bit convoluted given the differ ways they were used) -
we instead just iterate over the collection of ids directly.

We still lean on normalizr schemas in a few places, as they're succinct
and reasonably easy to understand, but use per-component schemas so that
we only pull out the objects we need.

Also refactor prop names so that they are all in camelCase.