Remove "unmet peer dependency" warnings
Also allows us to pin @babel/core to an up-to-date version. Remaining
warnings are from @types/react, but this is fixed in the versions of the
packages we're pulling in (via `peerDependenciesMeta.optional`) -
however the version of Yarn we currently use in the Docker image is too
older than the introduction of that field.