Make OctokitClient methods standout
Previously, with a method like `branches` it was unclear as to whether
the method was querying the database or calling the GH API. By prefixing
Octokit methods with `octokit_` we make this clearer.

For maintaining the existing interface (mostly in Repo), we add an
option to make non-prefixed method aliases.