Add sliding side panel to <TicketModal />
For editing info in the sidebar, means that we don't need to juggle
extra popovers/modals on an existing modal interface, and gives us a bit
more space for combobox inputs than we might have had otherwise.

Using Reach Router for the individual edit fields, since it's relatively
simple (we'd otherwise need to keep some state locally and drill it down
to child components, or keep some state in the redux store - and what is
the URL but a global store, with some existing dispatch mechanisms (i.e.
<Link />)?