Add LabellingsController
For updating labels on a ticket. Acts via a set of add/remove label
names, since we don't want to blindly clobber the set of labels on
GitHub as they may have changed without our realising (we also would
need to account for the "status: <swimlane>" labels).

Since the GitHub API returns the up-to-date set of labels on an issue
for the endpoints we use, we store this set on every API call, and only
update the DB once we've made all the calls (this also means that we
don't need to check if we're only doing additions or deletions).