Update server when "Save" clicked on <LabelPicker />
Uses optimistic update to immediately reflects changes in the UI, then
updates labels in the Redux store using the server response when it
arrives. If the server returns an error, we revert our changes (and
hope that something else hasn't made changes in the interim).

TODO: notification if error returned from server.