Add component for displaying react-toastify messages
We skip the bundled CSS for our own styling via Tailwind CSS, and wrap
the <ToastContainer /> in a <@reach/portal /> so that we can render them
at the end of the document and thus be reasonably sure that they will
appear above all over components (<Header />, <Modal /> ...).

Since `cssTransition` uses `onAnimationEnd` we can't use Tailwind
transform classes as we do elsewhere, so instead we implement our own
animation based on the `Slide` component from react-toastify, but
translating by a fixed width rather than a percentage for (hopefully)
smoother transitions.