Fix <TicketSidebar /> not being hidden completely when slide out
If there's no comments on the ticket and/or the ticket has a short body,
the <TicketFeed /> doesn't extend far enough down the y-axis to cover
the translated sidebar. We fix this by instead translating the sidebar
component itself, and putting `overflow: hidden` on the previously-
translated `position: sticky` wrapper.