Port `BOARD_LOAD` action to RTK
Rather than handling the action in each slice reducer (which is
repetitive and tedious), we just handle it in one place (since the data
is uniform) in a separate reducer, which then handles off the result to
the slice reducers for further manipulation if needed.

This means that the rote upserting of db data can be done simply in one
place, leaving the slice reducers to handle the exceptions. For
example, moving a ticket returns that ticket with its linked PRs, repo,
ticket etc. - so the upserting of that info can be done in one place,
with the 'custom' changes - moving the ticket from one swimlane to
another - can be handled by the `swimlanes` slice.

Amends the following helpers:

- `upsert` can now handle normalized objects (i.e. keyed by id).
- `createRequestThunk` can now take a `process` function to post-process
  the API response (e.g. normalisation).