Port 'load full ticket' actions to RTK
Rather than relying on `current.boardTicket`, we now set the id via the
location state (for a consistent interface) once we have it, which does
mean that `connect` on <TicketFeed /> can no longer assume that the `id`
prop does not change.

Note that we need to wait for `navigate` to complete before disabling
the 'loading' mode to ensure that the `id` is present in the location
state.