Refactor <TicketModal /> use of `id`
Fetching the id from the location state doesn't work if we navigate to
a sub-page of the modal, so we instead switch to `useSelector` so that
we can close over the id we capture in local state and use that instead.

We now use location state as a 'bonus' to set the initial value of the
local state, but otherwise we need to wait until the API request
completes.

Since we don't use an `id` prop, we no longer need a wrapper component
for the router and can use <TicketModal /> directly.

Also extract section + previous section logic into its own hook.

TODO: handle API errors from request - error page?