Add JS skeleton for component demos
Using same dependencies at the app. Tried out Storybook, but it pulls in
_a lot_ of transitive dependencies and didn't manage to build with the
current Node/Yarn combination we use. Seeing as I want a way to display
lots of different combinations of the same component on the same page as
a way to do visual regression testing, we might as well write our own
thing, since we'd probably end up doing something similar with Storybook
anyway.