Fix scroll resetting when closing ticket modal
Previously, when closing a ticket modal the horizontal scroll position
of the board would reset to zero (i.e. all the way back to the left-hand
side).

Looks like this is due to Reach Router's focus management and something
in Chrome, so this stops it managing focus (pretty sure @reach-ui/dialog
already does something to handle focus anyway, since it traps focus
within the modal).