Add debugging for number-less Tickets
I suspect that GitHub is sending us webhooks for PRs under the `issue`
event, and/or we're importing PRs as issues from the GitHub API, which
sometimes mixes in PRs with issues in the `/org/repo/issues` endpoint.

This addition should give us more info and allow us to track down the
source of the problem.