Use Repo object rather than re-finding for Ticket/PullRequest.import
Brings those models inline with the rest of the app. Since we've already
retrieved a Repo record in all cases that the `.import` method is used, there's
no point finding it again.

This also means that the `.find_by_remote` methods are redundant and are thus
removed.

Suspect this fixes a potential bug in PullRequestRefresher, since we were
already passing in a Repo object to `PullRequest.import`, which would have been
using the wrong `id`, but haven't tested this theory.