Add service for building aggregated correlated subqueries
In order to load data in an efficient way in the normalised format that
we use in Redux, we need a way to load the ids of an association in the
same query that we use to load the parent models. This can be done in a
few ways, but correlated subqueries offer a relatively easy way to do
this, and allow us to customise the association being loaded (e.g. only
loading the first 10 tickets in a swimlane, for all swimlanes on a
board) by just writing a 'standard' query (and not windowing or
similar).

We use the association 'reader' that Rails uses as an easy way to build
up the subquery, then walk the Arel AST to swap out the bind variable
that points to the model in favour of a reference to the relevant
foreign key in the parent query. We then do a lateral join (a `CROSS
JOIN` in this case is the same as `LEFT JOIN LATERAL ... ON TRUE`), then
aggregate the subquery into an array that can then be referenced by the
parent query.