Add service to efficiently construct Redux JSON
Takes a 'plan' that lists the models to be loaded, along with their
associations, then iterates through the models, first loading the
'principal' models, then collecting the ids from the specified
associations and uses those to fetch the second 'level' of models, and
so on until we've loaded all the models referenced.

The collection of loaded records is structured the same as the Redux
store - a Hash keyed by record classes, with the values being Hashes of
individual records, keyed by their id.

The way the plan is looped over means that it is order-dependent (we
don't do any topological sort of dependencies first), but should work
so long as a 'top-level' reference to a model is placed *after* any
references to that model as an association.