Fix RSpec not using Flipper memory adapter
Metadata filters only work if the metadata key is present on an spec,
which isn't the case for non-standard-Rails entries under `app/` (in
this specific case, `app/rules`), since it seems RSpec doesn't infer any
type.

Also wrap memory adapter with instrumented adapter, so we get better
debugging info in the test log.