Fix rules firing on the wrong event
By (ab)using ActiveSupport::Callbacks, we were skipping running the
trigger block if any of the `if` conditions returned false (the event
matching particularly), and thus we weren't actually halting the
callback chain before running `call`.

Since triggers are intended to be "any-of", we instead forego using
callbacks directly, and instead store them separately and check whether
any match as a single step in the callback chain, which ensures that we
run all the checks we need to, and thus are able to halt the callback
chain correctly.