Split out trigger checks for rules
Means that we don't load the board/repo (for `enabled_for_board?`) for
rules that have been triggered for events that they don't handle - this
is because we only scope (with Wisper) by class, and so rules are still
run for events they have no trigger for.

By splitting out the trigger check to 1) an attributes check and 2) a
callback check, we can do the attributes check earlier in the callback
chain, and skip checking whether there's an associated board rule
enabled. This is more efficient (and hopefully) quicker, since we can
call the attributes check only using data we already have in memory,
without having to touch the db for the board rule rule - or the trigger
callback.