Fix no-longer-supported parameterised scope
Rails now throws an error if we `includes(:head_commit_statuses)`. Since
we don't eager-load that scope anywhere, we just remove that part of the
scope, and add the additional `repo_id` foreign key part to the instance
method where it's used.

Since we don't have the switch between eager-loading and instance
method, we delete the tests for this as well since we no longer need
them.