Proxy GitHub images via our server
GitHub now requires authentication for images recently uploaded to
private repos and embedded in Markdown, which means we can't load them
on FlightPlan's domain as the user's GitHub session cookie(s) won't be
sent.

Fix by proxying the images via our server, where we can fetch the images
(or other assets) using the user's OAuth token. At the moment, GitHub
redirects to a signed S3 URL with 5 mins expiry, and so it's sufficient
to just hand back the redirect to the browser to follow (rather than
downloading and serving the image ourselves).

No tests as this is very much dependent on GitHub's implementation, and
so any tests we would write would not catch regressions caused by them
changing their implementation.